Separated api tests from selenium tests in different files
