+++ jobfindr/tests_api.py
@@ -1,0 +1,11 @@
+from django.test import TestCase, Client
+
+DOMAIN = 'host.docker.internal'
+CONTAINER_URL = f'http://{DOMAIN}:3000'
+
+# Create your tests here.
+class IndexTests(TestCase):
+    def test_index(self):
+        c = Client()
+        response = c.get("/api/")
+        self.assertEqual(response.status_code, 200)
